factor out common code in unordered_map, set
This also gives the pmh algorithm tables an explicit name, and
replaces the tuple with a struct so that the membes have labels.
This may improve readability somewhat.

It also makes the hasher belong to the pmh_tables rather than the
container, since it sort of becomes an implementation detail of the
pmh hash function that is produced.