add maybe_seed structure
this means we don't use the sign bit anymore to indicate
whether the first table entry is an index or a seed

this creates some wasted memory but it means the hash seeds
can be chosen by a PRG

there's also a subtle optimization in here: when the bucket
is a singleton, we just put the index to that item from the
items array into the *first* hash table,
we don't bother putting it into the second hash table. this
skips a level of indirection on such queries, since we don't
even have to look at the second hash table then, and we
can delete a bunch of code in the pmh function associated to
the "freelist" also.