multiply_shift hashing, and seeded initial hash with retries
This commit introduces a hash-function adapter which gets combined
with the user-specified hasher.

The user-specified hasher is no longer used with a seed after this
commit, instead, the user hasher gets wrapped in a multiply-shift
hasher. The multiply shift hasher hashes the input with the user
hasher, then it multiplies the resulting uint64_t by the seed.
then it shifts down the most significant bits that we care about.

Multiply-shift hashing is "linear", but when the seed is
pseudorandom and we can also try multiple seeds at compile-time,
it should produce great statistical properties without a run-time
penalty.

We also use seeds in the "initial" hash in the pmh table. we adopt
a scoring system, where we add up the sum of the squares of the
bucket sizes, and if they exceed 10 * N, we try another seed.

This may increase the probability of success of the algorithm on
pathological inputs. For instance, suppose we want to store the
numbers 1, 17, 33, 49, 65, in a frozen::set. Before this commit,
the first hash, elsa with no seed, is the identity map, and then
we mod by M, which would be 8 here. Since all these numbers are
congruent mod 8, they all get hashed to the same bucket in the
first round. Then we have to find a single seed that hashes all
of them to different buckets in the second phase -- essentially,
the algorithm has "degenerated" to just randomly guessing a seed,
and this will happen whenever all or most of the inputs go to the
same bucket.

After this commit, the guarantee is as follows: As long as elsa
manages to send each input to a different uint64_t, we should
be able to find by trial and error a good seed that breaks them
up across many buckets in the first phase, and then good seeds
that manage to place each of those (relatively small) buckets.

It's important to note that if for two different items, elsa
sends them to the same 64 bit integer, this algorithm will
always fail. (Becase we no longer use the seeded version of elsa.)

This is not all that different from the previous situation --
if the compiler reports an infinite constexpr loop, the user needs
to manually change the hash function somehow, either changing the
seeded part or the unseeded part.

It might be a drawback when hashing "large" objects. For instance,
if we are hashing frozen::string, the elsa implementation that
takes a seed uses the seed together with all of the characters of
the string to produce the hash. So for any two strings that are
different, there is going to be a seed such that their hash is
different. In the current implementation, only the unseeded version
of elsa hash for string is used, and then the seed is incorporated
only after that 64 bit hash is used. So, if there are two strings
whose 64 bit hash is identical, the algorithm will always fail
because the multiply_shift hasher can never find a seed that
separates them.

One possible way to fix that would be to use SFINAE when we are
adapting the hash -- if the user has specified a "seeded" hash also,
then use their implementation, but if not, then use multiply_shift.

A nice feature of the muliply shift hashing is that it is extremely
fast -- we are now able to use the seeded hash twice, once in the
first table and once in the second table, and still be several times
faster than the std::unordered_map, on all targetted compilers.