use multiply_shift only when elsa does not provide seeded hash
this commit uses SFINAE on the user-provided hash

if it does not provide a "seeded" version of the hash, then we
make one on the spot using multiply-shift hashing.

if it does then we defer to their seeded hash function

this also comments out the default elsa hash function for integral
types, so that we use multiply-shift hashing for those, which
I think is slightly faster.