pmh: impose a bucket size maximum, and do retries until we reach it
we also add a test case that was failing previously when all the
items end up in the same bucket.

this should improve the worst case runtime of the pmh routine
significantly, and prevent some timeouts. it should also reduce the
memory usage of the pmh routine

the retry part of this was part of commit 497252f1a3