Fix control flow around bucket overflows in make_pmh_buckets
This is a correctness issue reported in frozen issue #95

The issue is that the `continue` is meant to continue the outer
`while (1)` loop, so that we wipe out the buckets and try another
seed. As written, the code will simply not place the item that
overflowed. Then lookup of that item will likely fail.

I have never seen this happen in practice, I think because we
make the buckets pretty large, like sqrt(N), during the construction.
So likely overflow never occurs.
Otherwise I would expect the map to give a junk answer in practice.

I think I introduced this bug in this commit:
https://github.com/serge-sans-paille/frozen/commit/2897706fc23936a04666e9e9e3d60903c7866c7b#diff-b7220304cb6955468f4a0071ade6af7e