feat: support storing strings contained in pairs & tuples too
This is necessary for storing strings in maps in a way that doesn't
require dynamic relocations when part of position-independent-code.

Using the tuple accessor (meta-) functions has the side benefit of this
being generic enough to be useful (without modifications) for something
like a multi_index type in the future.