fix(map): avoid returning references to temporaries
When our underlying iterator is a proxy iterator it's dereferencing
operator (`*`) will return a temporary. The .at() method should *not*
return a reference to it or one of its members (->second). Instead it's
`second` member should be returned as is. A value when it's a value
type, a reference when it's a reference type.