perf: use smallest possible integer to encode sizes and offsets
For large string tables (large number of elements) with relatively short
strings this has a huge effect. Easily cutting such a string table in
half.

For unordered maps/sets this is exclusively a size benefit. For ordered
maps/sets this significantly increases the probability of L1 cache hits
and thus also provides a potential performance benefit.