Removed prefix from code so it can be copied and executed
If there are > in front of every line the terminal will complain. This is not a problem if you copy line by line and can just exclude them. But modern terminals allow pasting multiple lines of instructions and execute them after each other. The > prevent simply copying all lines and pasting them like with the copy code button